Fix bug on mac where casting didn't work in calc_cellareas() (lat_bounds)

+++ classes/cccmautils.py
@@ -415,8 +415,8 @@ def calc_cellareas(lat,lon, repeat=None):
 #    lat_bounds = np.array([(lat[1:nlat] + lat[0:(nlat-1)]) / 2])
     #print lat_bounds # @@
 
-    lat_bounds = np.insert(lat_bounds,0,-90)
-    lat_bounds = np.append(lat_bounds,90)
+    lat_bounds = np.insert(lat_bounds,0,np.float32(-90))
+    lat_bounds = np.append(lat_bounds,np.float32(90))
 
     del_phi = np.sin(lat_bounds[1:]*deg2rad) - np.sin(lat_bounds[0:-1]*deg2rad)
     #print del_phi # @@
